# services/gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OlimpicApp — Tournament Platform API
  description: |
    Multi-sport tournament management platform.
    All endpoints require JWT authentication via httpOnly cookies unless marked as public.
    The gateway proxies requests to internal microservices.
  version: 1.0.0
  contact:
    name: Tournament Platform Team
servers:
  - url: http://localhost:3000
    description: Local development gateway
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Announcement:
      type: object
      properties:
        id: { type: string, format: uuid }
        tournamentId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        title: { type: string }
        content: { type: string }
        priority: { type: string, enum: [low, normal, high, urgent] }
        isPinned: { type: boolean }
        publishedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
...
    GalleryPhoto:
      type: object
      properties:
        id: { type: string, format: uuid }
        tournamentId: { type: string, format: uuid, nullable: true }
        matchId: { type: string, format: uuid, nullable: true }
        teamId: { type: string, format: uuid, nullable: true }
        url: { type: string }
        thumbnailUrl: { type: string, nullable: true }
        caption: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
security:
  - cookieAuth: []
paths:
...
  /api/announcements:
    get:
      tags: [Announcements]
      summary: List announcements
      parameters:
        - name: tournamentId
          in: query
          schema: { type: string, format: uuid }
        - name: priority
          in: query
          schema: { type: string, enum: [low, normal, high, urgent] }
      responses:
        '200':
          description: Announcement list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Announcement' }
    post:
      tags: [Announcements]
      summary: Create announcement (organizer+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tournamentId, title, content]
              properties:
                tournamentId: { type: string, format: uuid }
                title: { type: string, minLength: 2, maxLength: 200 }
                content: { type: string }
                priority: { type: string, enum: [low, normal, high, urgent] }
                isPinned: { type: boolean }
      responses:
        '201': { description: Announcement created }
  /api/announcements/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Announcements]
      summary: Get announcement by ID
      responses:
        '200': { description: Announcement details }
    put:
      tags: [Announcements]
      summary: Update announcement (organizer+)
      responses:
        '200': { description: Updated }
    delete:
      tags: [Announcements]
      summary: Delete announcement (admin only)
      responses:
        '204': { description: Deleted }

  # ── Payments ─────────────────────────────────────
...
  /api/gallery:
    get:
      tags: [Gallery]
      summary: List gallery photos
      parameters:
        - name: tournamentId
          in: query
          schema: { type: string, format: uuid }
        - name: matchId
          in: query
          schema: { type: string, format: uuid }
        - name: teamId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Photo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/GalleryPhoto' }
    post:
      tags: [Gallery]
      summary: Upload photo metadata (organizer+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                tournamentId: { type: string, format: uuid, nullable: true }
                matchId: { type: string, format: uuid, nullable: true }
                teamId: { type: string, format: uuid, nullable: true }
                url: { type: string, format: uri }
                thumbnailUrl: { type: string, format: uri, nullable: true }
                caption: { type: string, maxLength: 500, nullable: true }
      responses:
        '201': { description: Photo record created }
  /api/gallery/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Gallery]
      summary: Get photo by ID
      responses:
        '200': { description: Photo details }
    delete:
      tags: [Gallery]
      summary: Delete photo (admin only)
      responses:
        '204': { description: Deleted }